ci: fix to pass build
